Handle symlinks, IO errors, and a cabal fix.
[`getSymbolicLinkStatus`][gls] calls [`lstat`][lstat] instead of
[`stat`][stat]. The difference is that [`lstat`][lstat] won’t
dereference a symbolic link, and will get the link itself. This
accomplishes two things:

1. It will return a [`FileStatus`][fst] block for the link rather than
   the referenced file.
2. If the link doesn’t reference an existing file, it won’t throw an
   [`IOError`][ioe].

By getting the [`FileStatus`][fst] of the link rather than the file it
references, we’ll actually recreate links on the destination, and if
the referenced file doesn’t exist, the program won’t terminate with an
unhandled [`IOError`][ioe] whilst trying to open a file it doesn’t
need to.

[gls]: _
http://hackage.haskell.org/package/unix-2.7.2.2/docs/System-Posix-Files.html#v:getSymbolicLinkStatus _
`getSymbolicLinkStatus`
[fst]: http://hackage.haskell.org/package/unix-2.7.2.2/docs/System-Posix-Files.html#t:FileStatus _
`FileStatus`
[ioe]: http://hackage.haskell.org/package/base-4.10.0.0/docs/System-IO-Error.html#t:IOError _ `IOError`

[lstat]: https://linux.die.net/man/3/lstat `lstat`
[stat]: https://linux.die.net/man/3/stat `stat`

Logging `IOError`s to `stderr` rather than stoppin is handy when
trying to move data off of a filesystem that, say, ecryptfs has
damaged after a Linux distribution’s own special kernel failed to
maintain the ability to mount a filesystem or even build the modules
that are compatible with their kernel in the first place.

Including `Paths_sync_mht` in `other-modules` sections for executables
will squash a Cabal warning, and also make sure that the module is
actually available when packaged.